data_loading_callbacks refactored and working!
