+++ code/python_tooling/dashboard/run.py
@@ -1,0 +1,4 @@
+from dashboard.app import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
